+++ src/config.ts
@@ -1,15 +1,14 @@
 import Utils from './utils';
 import Validators from './validators';
-import { ClientOptions } from './client';
-import { Logger, type LoggerOptions } from './logger';
 import { SparkError } from './error';
 import { Authorization } from './auth';
 import { Interceptor } from './http';
+import { ClientOptions } from './client';
+import { Logger, LoggerOptions } from './logger';
 import { DEFAULT_MAX_RETRIES, DEFAULT_TIMEOUT_IN_MS, ENV_VARS } from './constants';
 
 export class Config {
   readonly #options!: string;
-  readonly #interceptors: Set<Interceptor> = new Set<Interceptor>();
 
   readonly baseUrl!: BaseUrl;
   readonly auth!: Authorization;
@@ -19,16 +18,17 @@ export class Config {
   readonly allowBrowser!: boolean;
   readonly logger!: LoggerOptions;
   readonly extraHeaders: Record<string, string> = {};
+  readonly interceptors: Set<Interceptor> = new Set<Interceptor>();
 
   constructor({
-    baseUrl = Utils.readEnv(ENV_VARS.BASE_URL),
+    baseUrl: url = Utils.readEnv(ENV_VARS.BASE_URL),
     apiKey = Utils.readEnv(ENV_VARS.API_KEY),
     token = Utils.readEnv(ENV_VARS.BEARER_TOKEN),
     ...options
   }: ClientOptions = {}) {
-    const numberValidator = Validators.positiveInteger;
+    const numberValidator = Validators.positiveInteger.getInstance();
 
-    this.baseUrl = BaseUrl.from({ url: baseUrl, tenant: options?.tenant, env: options?.env });
+    this.baseUrl = url instanceof BaseUrl ? url : BaseUrl.from({ url, tenant: options?.tenant, env: options?.env });
     this.auth = Authorization.from({ apiKey, token, oauth: options?.oauth });
     this.timeout = numberValidator.isValid(options.timeout) ? options.timeout! : DEFAULT_TIMEOUT_IN_MS;
     this.maxRetries = numberValidator.isValid(options.maxRetries) ? options.maxRetries! : DEFAULT_MAX_RETRIES;
@@ -58,45 +58,28 @@ export class Config {
     }
   }
 
-  get interceptors(): Interceptor[] {
-    return Array.from(this.#interceptors);
-  }
-
   get hasInterceptors(): boolean {
-    return this.#interceptors.size > 0;
+    return this.interceptors.size > 0;
   }
 
   get hasHeaders(): boolean {
     return !Utils.isEmptyObject(this.extraHeaders);
   }
 
-  /**
-   * Adds interceptors to the configuration (experimental feature).
-   * @param interceptors - methods to intercept requests and responses
-   */
-  addInterceptors(...interceptors: Interceptor[]): void {
-    interceptors.forEach((interceptor) => {
-      if (this.#interceptors.has(interceptor)) return;
-      this.#interceptors.add(interceptor);
-    });
-  }
-
-  addHeaders(headers: Record<string, string>): void {
-    Object.entries(headers).forEach(([key, value]) => {
-      this.extraHeaders[key] = value;
-    });
-  }
-
   copyWith(options: ClientOptions = {}): Config {
+    const { baseUrl: url, tenant, env = this.environment } = options;
     return new Config({
-      baseUrl: options.baseUrl || this.baseUrl.full,
-      apiKey: options.apiKey || this.auth.apiKey,
-      token: options.token || this.auth.token,
-      oauth: options.oauth || this.auth.oauth,
-      timeout: options.timeout || this.timeout,
-      maxRetries: options.maxRetries || this.maxRetries,
-      allowBrowser: options.allowBrowser || this.allowBrowser,
-      env: options.env || this.environment,
+      baseUrl:
+        url instanceof BaseUrl
+          ? url
+          : BaseUrl.from({ url: url ?? this.baseUrl.value, tenant: tenant ?? this.baseUrl.tenant, env }),
+      apiKey: options.apiKey ?? this.auth.apiKey,
+      token: options.token ?? this.auth.token,
+      oauth: options.oauth ?? this.auth.oauth,
+      timeout: options.timeout ?? this.timeout,
+      maxRetries: options.maxRetries ?? this.maxRetries,
+      allowBrowser: options.allowBrowser ?? this.allowBrowser,
+      logger: options.logger ?? this.logger,
     });
   }
 
@@ -108,16 +91,25 @@ export class Config {
 export class BaseUrl {
   readonly url!: URL;
 
-  private constructor(
+  protected constructor(
     baseUrl: string,
     readonly tenant: string,
   ) {
     this.url = new URL(baseUrl + '/' + tenant);
   }
 
+  /**
+   * Builds a base URL from the given parameters.
+   * @param {object} options - the distinct parameters to build a base URL from.
+   * @param {string} options.url - the base URL to use.
+   * @param {string} options.tenant - the tenant name.
+   * @param {string} options.env - the environment name to use.
+   * @returns a BaseUrl
+   * @throws {SparkError} if a base URL cannot be built from the given parameters.
+   */
   static from(options: { url?: string; tenant?: string; env?: string } = {}): BaseUrl {
-    const stringValidator = Validators.emptyString;
-    const urlValidator = Validators.baseUrl;
+    const stringValidator = Validators.emptyString.getInstance();
+    const urlValidator = Validators.baseUrl.getInstance();
 
     if (urlValidator.isValid(options?.url)) {
       const url = new URL(options.url!);
@@ -126,12 +118,14 @@ export class BaseUrl {
       if (stringValidator.isValid(tenant, 'tenant name is required')) {
         return new this(url.origin, tenant!);
       }
-    }
-
-    if (options?.env && options?.tenant) {
+    } else if (options?.env && options?.tenant) {
       const env = options.env.trim().toLowerCase();
       const tenant = options.tenant.trim().toLowerCase();
       return new this(`https://excel.${env}.coherent.global`, tenant);
+    } else {
+      // capture errors for missing parameters
+      stringValidator.isValid(options?.env, 'environment name is missing') &&
+        stringValidator.isValid(options?.tenant, 'tenant name is missing');
     }
 
     const errors = urlValidator.errors.concat(stringValidator.errors);
@@ -141,14 +135,17 @@ export class BaseUrl {
     });
   }
 
+  /** The base URL */
   get value(): string {
     return this.url.origin;
   }
 
+  /** The base URL including the tenant name. */
   get full(): string {
     return this.url.toString();
   }
 
+  /** The base URL for the keycloak service. */
   get oauth2(): string {
     return this.to('keycloak').concat('/auth/realms/', this.tenant);
   }