+++ src/constants.ts
@@ -1,5 +1,6 @@
 export const DEFAULT_MAX_RETRIES = 2;
 export const DEFAULT_TIMEOUT_IN_MS = 60000; // 60 seconds
+export const RETRY_RANDOMIZATION_FACTOR = 1.5;
 export const SPARK_SDK = 'Spark JS SDK';
 
 export const ENV_VARS = {